refactor plan scan files
