+++ ams/server/src/main/java/com/netease/arctic/server/optimizing/plan/AbstractPartitionPlan.java
@@ -1,12 +1,29 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.netease.arctic.server.optimizing.plan;
 
+import com.netease.arctic.data.IcebergContentFile;
 import com.netease.arctic.data.IcebergDataFile;
 import com.netease.arctic.server.optimizing.OptimizingConfig;
 import com.netease.arctic.server.optimizing.OptimizingType;
 import com.netease.arctic.server.table.TableRuntime;
 import com.netease.arctic.table.ArcticTable;
-import org.apache.iceberg.DataFile;
-import org.apache.iceberg.DeleteFile;
 import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
 
 import java.util.List;
@@ -22,26 +39,24 @@ public abstract class AbstractPartitionPlan {
   protected ArcticTable tableObject;
   private long fromSequence = INVALID_SEQUENCE;
   private long toSequence = INVALID_SEQUENCE;
-  protected final long currentTime;
+  protected final long planTime;
   
   protected boolean canAddFile = true;
 
-  public AbstractPartitionPlan(TableRuntime tableRuntime, ArcticTable table, String partition) {
+  public AbstractPartitionPlan(TableRuntime tableRuntime, ArcticTable table, String partition, long planTime) {
     this.tableObject = table;
     this.partition = partition;
     this.config = tableRuntime.getOptimizingConfig();
     this.tableRuntime = tableRuntime;
     this.fragmentSize = config.getTargetSize() / config.getFragmentRatio();
-    this.currentTime = System.currentTimeMillis();
+    this.planTime = planTime;
   }
 
   public String getPartition() {
     return partition;
   }
 
-  public abstract void addFile(DataFile dataFile, List<DeleteFile> deletes);
-
-  public abstract void addFile(DataFile dataFile, List<DeleteFile> deletes, List<IcebergDataFile> changeDeletes);
+  public abstract void addFile(IcebergDataFile dataFile, List<IcebergContentFile<?>> deletes);
 
   public abstract boolean isNecessary();
